refactor(edge): isolate bootstrap response writeback and streaming flow
