+++ crates/edge/src/quic_listener/bootstrap/request.rs
@@ -44,10 +44,9 @@ use super::{
             AdmissionPolicyDecision, admission_rejection_response,
             evaluate_forwarding_pre_admission_policy,
         },
-        bootstrap_tls::{BootstrapStreamingBody, boxed_full},
         forwarding::BootstrapResolutionInput,
     },
-    bootstrap_error_response,
+    BootstrapStreamingBody, bootstrap_error_response, boxed_full,
     intake::BootstrapRequestIntake,
 };
