+++ crates/edge/src/quic_listener/bootstrap_tls.rs
@@ -1,41 +1,33 @@
 use std::{
     convert::Infallible,
-    pin::Pin,
     sync::{
-        Arc, RwLock,
+        Arc,
         atomic::{AtomicBool, AtomicUsize, Ordering},
     },
-    task::{Context, Poll},
     time::{Duration, Instant},
 };
 
 use bytes::Bytes;
 use http::{Request, Response, StatusCode};
-use http_body_util::{BodyExt, Full};
 use hyper::{
-    body::{Body, Frame, Incoming},
+    body::Incoming,
     server::conn::{http1, http2},
     service::service_fn,
-    upgrade,
 };
 use hyper_util::rt::TokioIo;
 use log::{debug, error, info, warn};
-use spooky_bridge::response::{
-    ResponseBodyMode, ResponseBodyPolicy, ResponseNormalizationInput,
-    ResponseNormalizationProtocol, ResponseProtocolConstraints, normalize_upstream_response,
-};
 use spooky_config::runtime::ListenerRuntimeConfig;
 use spooky_errors::ProxyError;
-use spooky_lb::upstream_pool::UpstreamPool;
 
 pub(super) use super::bootstrap::{BootstrapConnectionState, BootstrapStartupState};
 use super::{
     QUICListener,
     bootstrap::{
         BootstrapBuildRequestInput, BootstrapDispatchInput, BootstrapPolicyEvaluationInput,
-        BootstrapPreparedRoute, BootstrapRequestIntake, build_bootstrap_upstream_request,
-        dispatch_bootstrap_upstream, evaluate_bootstrap_request_policy,
-        prepare_bootstrap_request_intake,
+        BootstrapPreparedRoute, BootstrapRequestIntake, BootstrapWritebackInput, boxed_full,
+        build_bootstrap_upstream_request, dispatch_bootstrap_upstream,
+        evaluate_bootstrap_request_policy, prepare_bootstrap_request_intake,
+        write_bootstrap_response,
     },
     bootstrap::{
         PreparedBootstrapListenerStartup, prepare_bootstrap_listener_startup,
@@ -49,22 +41,17 @@ use crate::{
         bundle::RuntimeBundleHandle,
         connection::{
             guardrails::{
-                BodyLimitKind, REQUEST_BODY_TOO_LARGE_BODY, RESPONSE_BODY_TOO_LARGE_BODY,
-                RequestBodyGuardrailConfig, RequestBodyGuardrailDecision,
-                RequestBodyGuardrailInput, ResponseBodyGuardrailConfig,
-                ResponseBodyGuardrailDecision, ResponseBodyGuardrailInput,
-                checked_request_body_ingress, checked_response_body_guardrails,
+                BodyLimitKind, REQUEST_BODY_TOO_LARGE_BODY, RequestBodyGuardrailConfig,
+                RequestBodyGuardrailDecision, RequestBodyGuardrailInput,
+                checked_request_body_ingress,
             },
-            outcome::{
-                OutcomeBackendTarget, OutcomeRouteTarget, observe_backend_response_status,
-                observe_proxy_error_outcome, observe_status_outcome,
-            },
+            outcome::{OutcomeBackendTarget, OutcomeRouteTarget, observe_proxy_error_outcome},
         },
         shared_state::SharedRuntimeState,
     },
 };
 
-type BootstrapServiceFuture = Pin<
+type BootstrapServiceFuture = std::pin::Pin<
     Box<
         dyn std::future::Future<
                 Output = Result<
@@ -76,154 +63,6 @@ type BootstrapServiceFuture = Pin<
             > + Send,
     >,
 >;
-
-pub(in crate::quic_listener) struct BootstrapStreamingBody {
-    inner: Incoming,
-    guardrails: Option<ResponseBodyGuardrailConfig>,
-    declared_content_length: Option<usize>,
-    bytes_seen: usize,
-    prebuffered_bytes: usize,
-    capped: bool,
-    backend_accounting: Option<BootstrapBackendAccounting>,
-}
-
-struct BootstrapBackendAccounting {
-    upstream_pool: Arc<RwLock<UpstreamPool>>,
-    backend_index: usize,
-    start: Instant,
-    status: Option<u16>,
-    finished: bool,
-}
-
-impl BootstrapStreamingBody {
-    pub(in crate::quic_listener) fn new(inner: Incoming) -> Self {
-        Self {
-            inner,
-            guardrails: None,
-            declared_content_length: None,
-            bytes_seen: 0,
-            prebuffered_bytes: 0,
-            capped: false,
-            backend_accounting: None,
-        }
-    }
-
-    fn with_response_guardrails(
-        inner: Incoming,
-        max_body_bytes: usize,
-        declared_content_length: Option<usize>,
-        upstream_pool: Arc<RwLock<UpstreamPool>>,
-        backend_index: usize,
-        start: Instant,
-        status: Option<u16>,
-    ) -> Self {
-        Self {
-            inner,
-            guardrails: Some(ResponseBodyGuardrailConfig {
-                idle_timeout: Duration::MAX,
-                total_timeout: Duration::MAX,
-                max_body_bytes,
-                unknown_length_prebuffer_bytes: max_body_bytes,
-                chunk_bytes: usize::MAX,
-            }),
-            declared_content_length,
-            bytes_seen: 0,
-            prebuffered_bytes: 0,
-            capped: false,
-            backend_accounting: Some(BootstrapBackendAccounting {
-                upstream_pool,
-                backend_index,
-                start,
-                status,
-                finished: false,
-            }),
-        }
-    }
-
-    fn finish_backend_accounting(&mut self) {
-        if let Some(accounting) = self.backend_accounting.as_mut() {
-            if accounting.finished {
-                return;
-            }
-            crate::runtime::connection::outcome::finish_backend_request_accounting(
-                crate::runtime::connection::outcome::BackendRequestFinishInput {
-                    upstream_pool: Some(&accounting.upstream_pool),
-                    backend_index: Some(accounting.backend_index),
-                    elapsed: accounting.start.elapsed(),
-                    status: accounting.status,
-                },
-            );
-            accounting.finished = true;
-        }
-    }
-}
-
-impl Body for BootstrapStreamingBody {
-    type Data = Bytes;
-    type Error = Infallible;
-
-    fn poll_frame(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Option<Result<Frame<Self::Data>, Self::Error>>> {
-        if self.capped {
-            return Poll::Ready(None);
-        }
-
-        match Pin::new(&mut self.inner).poll_frame(cx) {
-            Poll::Ready(Some(Ok(frame))) => {
-                if let Some(guardrails) = self.guardrails
-                    && let Some(data) = frame.data_ref()
-                {
-                    if let Ok(next_state) = checked_response_body_guardrails(
-                        guardrails,
-                        ResponseBodyGuardrailInput {
-                            elapsed: Duration::ZERO,
-                            idle_for: Duration::ZERO,
-                            bytes_received: self.bytes_seen,
-                            prebuffered_bytes: self.prebuffered_bytes,
-                            next_chunk_bytes: data.len(),
-                            declared_content_length: self.declared_content_length,
-                            headers_emitted: true,
-                            progressive_emission_allowed: true,
-                            body_forwarding_enabled: true,
-                            exempt_from_body_size_cap: false,
-                        },
-                    ) {
-                        self.bytes_seen = next_state.next_state.bytes_received;
-                        self.prebuffered_bytes = next_state.next_state.prebuffered_bytes;
-                    } else {
-                        self.capped = true;
-                        self.finish_backend_accounting();
-                        return Poll::Ready(None);
-                    }
-                }
-                Poll::Ready(Some(Ok(frame)))
-            }
-            Poll::Ready(Some(Err(_))) => {
-                self.finish_backend_accounting();
-                Poll::Ready(None)
-            }
-            Poll::Ready(None) => {
-                self.finish_backend_accounting();
-                Poll::Ready(None)
-            }
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-impl Drop for BootstrapStreamingBody {
-    fn drop(&mut self) {
-        self.finish_backend_accounting();
-    }
-}
-
-pub(in crate::quic_listener) fn boxed_full(
-    body: Bytes,
-) -> http_body_util::combinators::BoxBody<Bytes, Infallible> {
-    Full::new(body).map_err(|never| match never {}).boxed()
-}
 
 fn bootstrap_route_target<'a>(route: &'a str) -> OutcomeRouteTarget<'a> {
     OutcomeRouteTarget { route }
@@ -612,7 +451,7 @@ impl QUICListener {
                                             }));
                                     }
                                 };
-                                let mut upstream_resp =
+                                let upstream_resp =
                                     match dispatch_bootstrap_upstream(BootstrapDispatchInput {
                                         upstream_req,
                                         prepared_route: &BootstrapPreparedRoute {
@@ -638,210 +477,24 @@ impl QUICListener {
                                         Err(response) => return Ok(response),
                                     };
 
-                                let status = upstream_resp.status();
-                                let normalized_response =
-                                    normalize_upstream_response(ResponseNormalizationInput {
-                                        upstream: spooky_bridge::response::UpstreamResponseView {
-                                            status,
-                                            headers: upstream_resp.headers(),
-                                            trailers: None,
-                                        },
-                                        body_mode: if suppress_downstream_body {
-                                            ResponseBodyMode::HeadRequest
-                                        } else {
-                                            ResponseBodyMode::Normal
-                                        },
-                                        constraints: ResponseProtocolConstraints {
-                                            protocol: ResponseNormalizationProtocol::Http1,
-                                            strip_connection_headers: true,
-                                            allow_trailers: false,
-                                            preserve_upgrade: is_websocket_upgrade
-                                                && status == StatusCode::SWITCHING_PROTOCOLS,
-                                        },
-                                    });
-                                let upstream_content_length = upstream_resp
-                                    .headers()
-                                    .get(http::header::CONTENT_LENGTH)
-                                    .and_then(|v| v.to_str().ok())
-                                    .and_then(|s| s.parse::<usize>().ok());
-                                let response_size_decision = checked_response_body_guardrails(
-                                    ResponseBodyGuardrailConfig {
-                                        idle_timeout: Duration::ZERO,
-                                        total_timeout: Duration::MAX,
-                                        max_body_bytes: max_response_body_bytes,
-                                        unknown_length_prebuffer_bytes: max_response_body_bytes,
-                                        chunk_bytes: 1,
-                                    },
-                                    ResponseBodyGuardrailInput {
-                                        elapsed: Duration::ZERO,
-                                        idle_for: Duration::ZERO,
-                                        bytes_received: 0,
-                                        prebuffered_bytes: 0,
-                                        next_chunk_bytes: 0,
-                                        declared_content_length: upstream_content_length,
-                                        headers_emitted: false,
-                                        progressive_emission_allowed: !normalized_response
-                                            .emission
-                                            .emit_end_stream_on_headers,
-                                        body_forwarding_enabled: matches!(
-                                            normalized_response.emission.body,
-                                            ResponseBodyPolicy::Forward
-                                        ),
-                                        exempt_from_body_size_cap: is_websocket_upgrade
-                                            && status == StatusCode::SWITCHING_PROTOCOLS,
-                                    },
-                                );
-                                if matches!(
-                                    response_size_decision,
-                                    Err(ResponseBodyGuardrailDecision::Reject {
-                                        kind: BodyLimitKind::BodySize,
-                                    })
-                                ) {
-                                    let _ = observe_proxy_error_outcome(
-                                        metrics.as_ref(),
-                                        bootstrap_route_target(&upstream_name),
-                                        Some(bootstrap_backend_target(
-                                            &upstream_name,
-                                            &backend_addr,
-                                            backend_index,
-                                        )),
-                                        request_start.elapsed(),
-                                        Some(StatusCode::SERVICE_UNAVAILABLE),
-                                        &ProxyError::Pool(
-                                            spooky_errors::PoolError::BackendOverloaded(
-                                                "response prebuffer cap".into(),
-                                            ),
-                                        ),
-                                        Some(crate::OverloadShedReason::ResponsePrebufferCap),
-                                    );
-                                    return Ok(Response::builder()
-                                        .status(StatusCode::SERVICE_UNAVAILABLE)
-                                        .header("alt-svc", &alt)
-                                        .body(boxed_full(Bytes::from_static(
-                                            RESPONSE_BODY_TOO_LARGE_BODY,
-                                        )))
-                                        .unwrap_or_else(|_| {
-                                            Response::new(boxed_full(Bytes::from_static(
-                                                b"error\n",
-                                            )))
-                                        }));
-                                }
-                                let _ = observe_status_outcome(
-                                    metrics.as_ref(),
-                                    bootstrap_route_target(&upstream_name),
-                                    Some(bootstrap_backend_target(
-                                        &upstream_name,
-                                        &backend_addr,
+                                write_bootstrap_response(BootstrapWritebackInput {
+                                    upstream_resp,
+                                    prepared_route: &BootstrapPreparedRoute {
+                                        endpoint: endpoint.clone(),
+                                        backend_addr: backend_addr.clone(),
                                         backend_index,
-                                    )),
-                                    request_start.elapsed(),
-                                    status,
-                                );
-                                if let Some(transition) = observe_backend_response_status(
-                                    crate::runtime::connection::outcome::BackendHealthObservationInput {
-                                        backend_addr: &backend_addr,
-                                        backend_index,
-                                        upstream_pool: Some(&upstream_pool),
-                                        status,
+                                        upstream_name: upstream_name.clone(),
+                                        upstream_policy: upstream_policy.clone(),
+                                        upstream_pool: Arc::clone(&upstream_pool),
                                     },
-                                ) {
-                                    crate::runtime::connection::outcome::log_backend_health_transition(
-                                        &backend_addr,
-                                        transition,
-                                    );
-                                }
-                                let mut resp_builder =
-                                    Response::builder().status(normalized_response.head.status);
-                                for header in &normalized_response.head.headers {
-                                    resp_builder = resp_builder.header(&header.name, &header.value);
-                                }
-                                resp_builder = resp_builder.header("alt-svc", &alt);
-                                if is_websocket_upgrade
-                                    && upstream_resp.status() == StatusCode::SWITCHING_PROTOCOLS
-                                {
-                                    let client_upgrade = match client_upgrade {
-                                        Some(u) => u,
-                                        None => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upgrade setup error\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-                                    let upstream_upgrade = upgrade::on(&mut upstream_resp);
-                                    tokio::spawn(async move {
-                                        let (client, upstream) = match tokio::try_join!(
-                                            client_upgrade,
-                                            upstream_upgrade
-                                        ) {
-                                            Ok(v) => v,
-                                            Err(err) => {
-                                                debug!(
-                                                    "Bootstrap WebSocket upgrade join failed: {}",
-                                                    err
-                                                );
-                                                return;
-                                            }
-                                        };
-                                        let mut client = TokioIo::new(client);
-                                        let mut upstream = TokioIo::new(upstream);
-                                        let _ = tokio::io::copy_bidirectional(
-                                            &mut client,
-                                            &mut upstream,
-                                        )
-                                        .await;
-                                    });
-                                    crate::runtime::connection::outcome::finish_backend_request_accounting(
-                                        crate::runtime::connection::outcome::BackendRequestFinishInput {
-                                            upstream_pool: Some(&upstream_pool),
-                                            backend_index: Some(backend_index),
-                                            elapsed: request_start.elapsed(),
-                                            status: Some(status.as_u16()),
-                                        },
-                                    );
-                                    return Ok(resp_builder
-                                        .body(boxed_full(Bytes::new()))
-                                        .unwrap_or_else(|_| {
-                                            Response::new(boxed_full(Bytes::new()))
-                                        }));
-                                }
-                                let resp_body = if matches!(
-                                    normalized_response.emission.body,
-                                    ResponseBodyPolicy::Suppress
-                                ) {
-                                    crate::runtime::connection::outcome::finish_backend_request_accounting(
-                                        crate::runtime::connection::outcome::BackendRequestFinishInput {
-                                            upstream_pool: Some(&upstream_pool),
-                                            backend_index: Some(backend_index),
-                                            elapsed: request_start.elapsed(),
-                                            status: Some(status.as_u16()),
-                                        },
-                                    );
-                                    boxed_full(Bytes::new())
-                                } else {
-                                    BootstrapStreamingBody::with_response_guardrails(
-                                        upstream_resp.into_body(),
-                                        max_response_body_bytes,
-                                        upstream_content_length,
-                                        Arc::clone(&upstream_pool),
-                                        backend_index,
-                                        request_start,
-                                        Some(status.as_u16()),
-                                    )
-                                    .map_err(|never| match never {})
-                                    .boxed()
-                                };
-
-                                Ok(resp_builder
-                                    .body(resp_body)
-                                    .unwrap_or_else(|_| Response::new(boxed_full(Bytes::new()))))
+                                    metrics: metrics.as_ref(),
+                                    request_start,
+                                    alt_svc: &alt,
+                                    suppress_downstream_body,
+                                    is_websocket_upgrade,
+                                    client_upgrade,
+                                    max_response_body_bytes,
+                                })
                             })
                         },
                     );
